+++ ListableUI/Sources/Internal/ItemCell.swift
@@ -208,7 +208,9 @@ extension UIView {
     
     fileprivate func recursiveAccessibleSubviews() -> [Any] {
         subviews.flatMap { subview -> [Any] in
-            if let accessibilityElements = subview.accessibilityElements {
+            if subview.accessibilityElementsHidden || subview.isHidden {
+                return []
+            } else if let accessibilityElements = subview.accessibilityElements {
                 return accessibilityElements
             } else if subview.isAccessibilityElement {
                 return [subview]