+++ ListableUI/Sources/Layout/Table/TableListLayout.swift
@@ -108,6 +108,9 @@ public struct TableAppearance : ListLayoutAppearance
     
     /// How the layout should flow, either horizontally or vertically.
     public var direction: LayoutDirection
+
+    /// If the list header should stick to the top when scrolled.
+    public var stickyListHeader: Bool
     
     /// If sticky section headers should be leveraged in the layout.
     public var stickySectionHeaders : Bool
@@ -154,6 +157,7 @@ public struct TableAppearance : ListLayoutAppearance
     /// Creates a new `TableAppearance` object.
     public init(
         direction : LayoutDirection = .vertical,
+        stickListHeader: Bool = false,
         stickySectionHeaders : Bool = true,
         pagingBehavior : ListPagingBehavior = .none,
         itemPositionGroupingHeight : CGFloat = 0.0,
@@ -163,6 +167,7 @@ public struct TableAppearance : ListLayoutAppearance
         layout : Layout = .init()
     ) {
         self.direction = direction
+        self.stickyListHeader = stickListHeader
         self.stickySectionHeaders = stickySectionHeaders
         self.pagingBehavior = pagingBehavior
         self.itemPositionGroupingHeight = itemPositionGroupingHeight